# Optimizing memory usage in exponential histogram to Sketch conversion