--- a/pkg/quantile/pool.go
+++ b/pkg/quantile/pool.go
@@ -7,6 +7,8 @@ package quantile
 
 import (
 	"sync"
+
+	"github.com/DataDog/sketches-go/ddsketch/store"
 )
 
 const (
@@ -64,4 +66,54 @@ func getOverflowList() []bin {
 
 func putOverflowList(a []bin) {
 	overflowListPool.Put(&a)
+}
+
+var floatKeyCountPool = sync.Pool{
+	New: func() interface{} {
+		a := make([]floatKeyCount, 0, defaultBinListSize)
+		return &a
+	},
+}
+
+func getFloatKeyCountList() []floatKeyCount {
+	a := *(floatKeyCountPool.Get().(*[]floatKeyCount))
+	return a[:0]
+}
+
+func putFloatKeyCountList(a []floatKeyCount) {
+	if cap(a) >= defaultBinListSize {
+		floatKeyCountPool.Put(&a)
+	}
+}
+
+var keyCountPool = sync.Pool{
+	New: func() interface{} {
+		a := make([]KeyCount, 0, defaultBinListSize)
+		return &a
+	},
+}
+
+func getKeyCountList() []KeyCount {
+	a := *(keyCountPool.Get().(*[]KeyCount))
+	return a[:0]
+}
+
+func putKeyCountList(a []KeyCount) {
+	if cap(a) >= defaultBinListSize {
+		keyCountPool.Put(&a)
+	}
+}
+
+var denseStorePool = sync.Pool{
+	New: func() interface{} {
+		return store.NewDenseStore()
+	},
+}
+
+func getDenseStore() store.Store {
+	return denseStorePool.Get().(store.Store)
+}
+
+func putDenseStore(s store.Store) {
+	denseStorePool.Put(s)
 }