### What does this PR do?

This PR introduces memory allocation optimizations in the DDSketch to Sketch conversion process, resulting in reduced GC pressure during high-throughput scenarios.

## Changes

1. **Object Pooling with `sync.Pool`**:
   - Added pools for `KeyCount` slices and `DenseStore` objects to reduce allocation overhead
   - Added pool for bin list ([]bin) (the main offender)
   - Implemented getter/putter functions to manage pooled resources

I reused the `pool.go` file here, let me know if this is okay or if I should keep the new pools in the same file.

3. **Build System Enhancements**:
   - Updated Makefile to use configurable path for golangci-lint
   - Added GOPATH and GOBIN variables for better environment compatibility

I can revert this change, but in my local machine I had problems where my local installation of golanglint-ci was taking precedence over the binary in GOBIN.

## Benefits

These changes significantly reduce memory allocations during DDSketch to Sketch conversion operations, particularly in high-throughput scenarios where many conversions happen in succession. By reusing slice and store objects, we avoid repeated allocations and subsequent garbage collection.

```
goos: darwin
goarch: arm64
pkg: github.com/DataDog/opentelemetry-mapping-go/pkg/quantile
cpu: Apple M3 Pro
                             │ /tmp/old-code.txt │        /tmp/v2-new-code.txt         │
                             │      sec/op       │    sec/op     vs base               │
DDSketchConversion/Small-11         43.10µ ± 13%   37.02µ ±  1%  -14.10% (p=0.002 n=6)
DDSketchConversion/Medium-11        45.46µ ±  8%   37.13µ ± 13%  -18.32% (p=0.002 n=6)
DDSketchConversion/Large-11         45.10µ ± 18%   37.06µ ±  2%  -17.83% (p=0.002 n=6)
geomean                             44.54µ         37.07µ        -16.77%

                             │ /tmp/old-code.txt │        /tmp/v2-new-code.txt         │
                             │       B/op        │     B/op      vs base               │
DDSketchConversion/Small-11        174.90Ki ± 0%   13.60Ki ± 0%  -92.22% (p=0.002 n=6)
DDSketchConversion/Medium-11       174.91Ki ± 0%   13.61Ki ± 1%  -92.22% (p=0.002 n=6)
DDSketchConversion/Large-11        174.93Ki ± 0%   13.61Ki ± 0%  -92.22% (p=0.002 n=6)
geomean                             174.9Ki        13.61Ki       -92.22%

                             │ /tmp/old-code.txt │       /tmp/v2-new-code.txt        │
                             │     allocs/op     │ allocs/op   vs base               │
DDSketchConversion/Small-11           33.00 ± 0%   37.00 ± 0%  +12.12% (p=0.002 n=6)
DDSketchConversion/Medium-11          33.00 ± 0%   37.00 ± 0%  +12.12% (p=0.002 n=6)
DDSketchConversion/Large-11           33.00 ± 0%   37.00 ± 0%  +12.12% (p=0.002 n=6)
geomean                               33.00        37.00       +12.12%

```

## Testing

All tests pass with the new implementation, confirming that the optimization maintains correctness while improving resource usage.


### Motivation

While analyzing memory profiles from our production application, we identified the DDSketch conversion path as a significant contributor to memory pressure. Profiling revealed that frequent allocations of KeyCount slices and DenseStore objects during high-throughput metric processing were causing increased GC activity and memory usage spikes. This optimization addresses these specific hotspots to improve overall application performance and resource utilization.

![image](https://github.com/user-attachments/assets/48bac1ab-3b40-4722-bc11-c64e616b7e73)