@pedro-stanaka 2025-04-22
> Thanks for the detailed description, could you add a changelog? Other than that LGTM

Thanks for the fast review 🙂 

Added feedback.